Add NetworkQueryRequestManager and StandaloneQueryRequestManager classes
