+++ packages/core/src/plugins/logStore/NetworkQueryRequestManager.ts
@@ -1,11 +1,7 @@
 import { MessageMetadata, StreamMessage } from '@logsn/client';
 import {
-	QueryFromOptions,
-	QueryLastOptions,
-	QueryRangeOptions,
 	QueryRequest,
 	QueryResponse,
-	QueryType,
 	SystemMessage,
 	SystemMessageType,
 } from '@logsn/protocol';
@@ -16,19 +12,14 @@ import { Readable } from 'stream';
 
 import { BroadbandPublisher } from '../../shared/BroadbandPublisher';
 import { BroadbandSubscriber } from '../../shared/BroadbandSubscriber';
-import {
-	LogStore,
-	MAX_SEQUENCE_NUMBER_VALUE,
-	MIN_SEQUENCE_NUMBER_VALUE,
-} from './LogStore';
+import { LogStore } from './LogStore';
 import { PropagationResolver } from './PropagationResolver';
 import { QueryResponseManager } from './QueryResponseManager';
+import { StandaloneQueryRequestManager } from './StandaloneQueryRequestManager';
 
 const logger = new Logger(module);
 
-export class QueryRequestManager {
-	private logStore?: LogStore;
-
+export class NetworkQueryRequestManager extends StandaloneQueryRequestManager {
 	constructor(
 		private readonly queryResponseManager: QueryResponseManager,
 		private readonly propagationResolver: PropagationResolver,
@@ -36,11 +27,11 @@ export class QueryRequestManager {
 		private readonly subscriber: BroadbandSubscriber
 	) {
 		//
+		super();
 	}
 
-	public async start(logStore: LogStore) {
-		this.logStore = logStore;
-
+	public override async start(logStore: LogStore) {
+		super.start(logStore);
 		await this.subscriber.subscribe(this.onMessage.bind(this));
 	}
 
@@ -80,57 +71,6 @@ export class QueryRequestManager {
 			this.propagationResolver.waitForPropagateResolution(queryRequest);
 		await this.publisher.publish(queryRequest.serialize());
 		return resolutionPromise;
-	}
-
-	public getDataForQueryRequest(queryRequest: QueryRequest) {
-		let readableStream: Readable;
-		switch (queryRequest.queryType) {
-			case QueryType.Last: {
-				const { last } = queryRequest.queryOptions as QueryLastOptions;
-
-				readableStream = this.logStore!.requestLast(
-					queryRequest.streamId,
-					queryRequest.partition,
-					last
-				);
-				break;
-			}
-			case QueryType.From: {
-				const { from, publisherId, limit } =
-					queryRequest.queryOptions as QueryFromOptions;
-
-				readableStream = this.logStore!.requestFrom(
-					queryRequest.streamId,
-					queryRequest.partition,
-					from.timestamp,
-					from.sequenceNumber || MIN_SEQUENCE_NUMBER_VALUE,
-					publisherId,
-					limit
-				);
-				break;
-			}
-			case QueryType.Range: {
-				const { from, publisherId, to, msgChainId, limit } =
-					queryRequest.queryOptions as QueryRangeOptions;
-
-				readableStream = this.logStore!.requestRange(
-					queryRequest.streamId,
-					queryRequest.partition,
-					from.timestamp,
-					from.sequenceNumber || MIN_SEQUENCE_NUMBER_VALUE,
-					to.timestamp,
-					to.sequenceNumber || MAX_SEQUENCE_NUMBER_VALUE,
-					publisherId,
-					msgChainId,
-					limit
-				);
-				break;
-			}
-			default:
-				throw new Error('Unknown QueryType');
-		}
-
-		return readableStream;
 	}
 
 	private async getHashMap(data: Readable) {
